requirements.txt
# Spot
bosdyn-client
bosdyn-mission
bosdyn-choreography-client
# Web app
flask
# Facial detection
opencv-python
psycopg2-binary
imgbeddings
torch
numpy
# Audio
playsound
